requirements.txt
pynput
python-dotenv
cryptography